requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
cfgv==3.3.1
distlib==0.3.3
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
filelock==3.4.0
identify==2.4.0
nodeenv==1.6.0
platformdirs==2.4.0
pre-commit==2.15.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.10.0
